Remove combine_for_fit rule
It was needed when I was using a fixed sample fraction and varied
reference fraction, but now I'm only using data splits where the
sample fraction is 1 minus the reference fraction.
This rule copied lots of files unnecessarily and ate up too much
storage space, even when marking the output files as temporary.